Simplified more code
